Create onion-bitfinex.xml
This leads to the onion site for Bitfinex. It is off, because it contains a self-signed https cert, so a user would have to manually confirm the security exception, which we shouldn't ask users to do. Another reason is that Bitfinex will not accept logins to the onion site unless the user explicitly allows it beforehand.